fix c-style error

+++ xla/service/gpu/runtime/thunk.cc
@@ -354,7 +354,7 @@ Thunk::ThunkInfo Thunk::ThunkInfo::WithProfileAnnotation(
   if (gpu_backend_config.ok()) {
     thunk_info.execution_stream_id =
         std::max<uint64_t>(kDefaultExecutionStreamId.value(),
-                           uint64_t(gpu_backend_config->operation_queue_id()));
+                           gpu_backend_config->operation_queue_id());
   }
   return thunk_info;
 }
